+++ Craftsman/Builders/ExtensionBuilders/ServiceConfigurationBuilder.cs
@@ -41,6 +41,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Resources;
 using Sieve.Services;
 using System.Reflection;
+using System.Text.Json;
+using System.Text.RegularExpressions;
 
 public static class {FileNames.WebAppServiceConfiguration()}
 {{
@@ -52,12 +54,8 @@ public static class {FileNames.WebAppServiceConfiguration()}
         builder.Services.OpenTelemetryRegistration(""{projectBaseName}"");
         builder.Services.AddInfrastructure(builder.Environment);
 
-        // using Newtonsoft.Json to support PATCH docs since System.Text.Json does not support them https://github.com/dotnet/aspnetcore/issues/24333
-        // if you are not using PatchDocs and would prefer to use System.Text.Json, you can remove The `AddNewtonSoftJson()` line
-        builder.Services.AddControllers(options => options.UseDateOnlyTimeOnlyStringConverters())
-            .AddJsonOptions(options => options.UseDateOnlyTimeOnlyStringConverters())
-            .AddJsonOptions(o => o.JsonSerializerOptions.ReferenceHandler = ReferenceHandler.IgnoreCycles)
-            .AddNewtonsoftJson();
+        builder.Services.AddControllers()
+            .AddJsonOptions(o => o.JsonSerializerOptions.ReferenceHandler = ReferenceHandler.IgnoreCycles);
         builder.Services.AddApiVersioningExtension();
 
         builder.Services.AddHttpContextAccessor();
@@ -67,7 +65,9 @@ public static class {FileNames.WebAppServiceConfiguration()}
         // registers all services that inherit from your base service interface - {boundaryServiceName}
         builder.Services.AddBoundaryServices(Assembly.GetExecutingAssembly());
 
-        builder.Services.AddMvc(options => options.Filters.Add<ErrorHandlerFilterAttribute>());
+        builder.Services
+            .AddMvc(options => options.Filters.Add<ErrorHandlerFilterAttribute>())
+            .AddJsonOptions(opt => opt.JsonSerializerOptions.AddDateOnlyConverters());
 
         if(builder.Environment.EnvironmentName != Consts.Testing.FunctionalTestingEnvName)
         {{
@@ -103,6 +103,62 @@ public static class {FileNames.WebAppServiceConfiguration()}
             }}
         }}
     }}
+}}
+
+// TODO these will be baked into System.Text.Json in .NET 7
+public static class DateOnlyConverterExtensions
+{{
+    public static void AddDateOnlyConverters(this JsonSerializerOptions options)
+    {{
+        options.Converters.Add(new DateOnlyConverter());
+        options.Converters.Add(new DateOnlyNullableConverter());
+    }}
+}}
+
+public class DateOnlyConverter : JsonConverter<DateOnly>
+{{
+    public override DateOnly Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    {{
+        if (reader.TryGetDateTime(out var dt))
+        {{
+            return DateOnly.FromDateTime(dt);
+        }};
+        var value = reader.GetString();
+        if (value == null)
+        {{
+            return default;
+        }}
+        var match = new Regex(""^(\\d\\d\\d\\d)-(\\d\\d)-(\\d\\d)(T|\\s|\\z)"").Match(value);
+        return match.Success
+            ? new DateOnly(int.Parse(match.Groups[1].Value), int.Parse(match.Groups[2].Value), int.Parse(match.Groups[3].Value))
+            : default;
+    }}
+
+    public override void Write(Utf8JsonWriter writer, DateOnly value, JsonSerializerOptions options)
+        => writer.WriteStringValue(value.ToString(""yyyy-MM-dd""));
+}}
+
+public class DateOnlyNullableConverter : JsonConverter<DateOnly?>
+{{
+    public override DateOnly? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    {{
+        if (reader.TryGetDateTime(out var dt))
+        {{
+            return DateOnly.FromDateTime(dt);
+        }};
+        var value = reader.GetString();
+        if (value == null)
+        {{
+            return default;
+        }}
+        var match = new Regex(""^(\\d\\d\\d\\d)-(\\d\\d)-(\\d\\d)(T|\\s|\\z)"").Match(value);
+        return match.Success
+            ? new DateOnly(int.Parse(match.Groups[1].Value), int.Parse(match.Groups[2].Value), int.Parse(match.Groups[3].Value))
+            : default;
+    }}
+
+    public override void Write(Utf8JsonWriter writer, DateOnly? value, JsonSerializerOptions options)
+        => writer.WriteStringValue(value?.ToString(""yyyy-MM-dd""));
 }}";
     }
 }